requirements.txt
openai
pushbullet.py
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
